Use a Consumer for processing the ResultSet
Continuing in my quest to use Java 8 features to its fullest, I am using
a Consumer in the processResult method. This would enable us to pass a
small piece of login to be executed for every ResultSet.

FetchResult class uses this method to print out the contents of the
table.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>