Process large dataset by using insensitive scroll
Using the earlier method, the JVM was getting killed by the kernel. This
led me to believe that this is a memory issue.

To optimize memory usage, the following three are essential.
a. autoCommit off.
b. Forward only scroll.
c. Reading the results in a read only fashion.
d. Using a small fetchsize. This would enable use of a cursor to cache
given number of rows per fetch.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>